@@ -16,12 +16,9 @@ final class ComplexityCollectionIterator implements Iterator
     /**
      * @psalm-var list<Complexity>
      */
-    private $items;
+    private array $items;
 
-    /**
-     * @var int
-     */
-    private $position = 0;
+    private int $position = 0;
 
     public function __construct(ComplexityCollection $items)
     {
